src/bridge/requirements.txt
Jinja2==2.7
PyYAML==3.10
boto==2.32.1
pyzmq==13.1.0
redis==2.10.3
requests==1.2.3
rethinkdb==1.13.0-3
Werkzeug==0.9.6
twilio>=3.6.0
guppy==0.1.10
